style: Apply code formatting by pylint
